Fix PortalService static method and WebhookService DurationMs cast
- PortalService: make GenerateJwtToken non-static (uses instance _portalSigningKey)
- WebhookService: cast DurationMs decimal to int for WebhookDeliveryResponse

+++ Accounting/Services/Implementations/WebhookService.cs
@@ -154,7 +154,7 @@ public class WebhookService : IWebhookService
             d.HttpStatusCode,
             d.IsSuccess,
             d.AttemptNumber,
-            d.DurationMs,
+            (int)d.DurationMs,
             d.DeliveredAt,
             d.ErrorMessage
         )).ToList();
